Move Player Challenge
The challenge was to move the player away from and towards the camera using the values in the inspector. I, instead, did this as a coding challenge and wrote a short script to move the player box away from the camera to a max distance, back to its original position and then stop.