Finished Obstacle Course
Added spinning obstacles and obstacles that drop on a timer. Gave the player a pink start location and end location. Made it so certain objects have collision and turn red when the player object collides with them. Added a script that keeps of a tally of the number of unique objects the player hits and prints the number to the console every time the player hits something new.